+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
 
+## patch
+* Set up CMS to handle new type: asset-teaser
+
+
 ## v0.1.0
 * initial version